split unit and integration tests in makefile
